apps/app_demo_page_ocr/pyproject.toml
[tool.poetry]
name = "app_demo_page_ocr"
version = "0.0.0"
description = ""
authors = [
           "Eric Harvey <eric.harvey@informed.iq>",
           "Rahul Salla <rahul.salla@informed.iq>"
           ]
readme = "README.md"
include = []
exclude = ["*/config.yaml", "*/event.json"]
packages = [{ include = "app_demo_page_ocr", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.7.4,<3.8"

[tool.poetry.group.layer.dependencies]
honeybadger = "0.8.0"
pyhumps = "1.6.1"
requests = "2.28.1"
boto3 = "1.24.86"
numpy = "1.19.1"
Pillow = "8.4.0"
scipy = "1.3.1"
opencv-python = "3.4.7.28"

[tool.poetry.group.docker.dependencies]
aws-lambda-powertools = "2.2.0"
fastjsonschema = "2.16.2"
boto3 = "1.24.86"


[tool.poetry.group.dev.dependencies]
pytest-srcpaths = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
pytest-xdist = "^2.5.0"
python-lambda = "^11.8.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.4.25"
flake8-import-order = "^0.18.1"
ipdb = "^0.13.9"
pycodestyle = "^2.4.0"
pytest-stub = "^1.1.0"
moto = "^4.0.6"
coverage = {extras = ["toml"], version = "^6.5.0"}
autopep8 = "^1.6.0"

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
  "focus: focus specific tests to run"
]

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 50

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
